# Conflicts:
#	csc207/src/main/java/com/labrats/app/App.java
#	csc207/src/main/java/interface_adapter/add_expense/AddExpenseController.java
#	csc207/src/main/java/use_case/add_expense/AddExpenseInputData.java

+++ csc207/src/main/java/entity/Transaction.java
@@ -11,12 +11,13 @@ public class Transaction {
     private String category;
     private Date date;
 
-    public Transaction(String name, int amount, String category, Date date) {
+    public Transaction(String name, double amount, String category, Date date) {
         this.name = name;
         this.amount = amount;
         this.category = category;
         this.date = date;
     }
+
 
     public String getName() {
         return name;
